requirements.txt
flask
requests
html2text
typer
rich
reportlab
Pillow